move tag compiler to view namespace
